testing/mochitest/tests/test_bug332848.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=332848
-->
<head>
  <title>Test for Bug 332848</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=332848">Mozilla Bug 332848</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">
<![CDATA[

/** Test for Bug 332848 **/

// parseChecker will become true if we keep parsing after calling close().
var parseChecker = false;

function test() {
  try {
    document.open();
    is(0, 1, "document.open succeeded");
  } catch (e) {
    is (e.code, DOMException.NOT_SUPPORTED_ERR,
        "Wrong exception from document.open");
  }

  try {
    document.write("aaa");
    is(0, 1, "document.write succeeded");
  } catch (e) {
    is (e.code, DOMException.NOT_SUPPORTED_ERR,
        "Wrong exception from document.write");
  }

  try {
    document.close();
    is(0, 1, "document.close succeeded");
  } catch (e) {
    is (e.code, DOMException.NOT_SUPPORTED_ERR,
        "Wrong exception from document.close");
  }
}

function loadTest() {
  is(parseChecker, true, "Parsing stopped");
  test();
  SimpleTest.finish();
}

window.onload = loadTest;

SimpleTest.waitForExplicitFinish();

test();
]]>
</script>
<script>
    parseChecker = true;
</script>
</pre>
</body>
</html>
